pledge/requirements.txt
Django
django-registration
-e hg+https://bitbucket.org/ubernostrum/django-registration#egg=django-registration
